Move keygen function to STcommon.py
